Mejorando estetica y visualizacion
